separate out tests for clarity
